AI_Business_Insights_Agent/requirements.txt
streamlit==1.28.0  
pandas==2.1.4  
sqlalchemy==2.0.25  
crewai==0.2.0  
crewai_tools==0.1.4  
python-dotenv==1.0.1  
